% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write_cb}
\alias{read_cb}
\alias{write_cb}
\title{Portable read/write access to system clipboard}
\usage{
write_cb(x, ...)

read_cb(...)
}
\arguments{
\item{x}{object to be written, it is coerced to character}

\item{...}{other arguments passed to/from other methods, including \code{condesc}
with connection description. See Details.}
}
\value{
For \code{write_cb} it is \code{NULL}.

For \code{read_cb}, it is a character vector with lines of content of the
clipboard.
}
\description{
These functions aim at providing portable way of reading and writing to
system clipboard. The primary motivation was to smooth the process of
copy-pasting computation results from R to Office programs (MSWord,
LibreOffice, etc.).
}
\details{
Interaction with system clipboard is very much system-dependent. The
particular method used depends on the type of the operating system, which is
queried from \code{Sys.info()["sysname"]}.

\code{write_cb} expects a single argument which is written into the
clipboard after being converted to character. Writing is performed via a
connection description of which can be specified with \code{condesc} or the
\code{clipboard.write} option.  Defaults are system-specific, see below.

On Windows writing to clipboard is performed via \code{file} connection.
Default description is \code{clipboard}.

On Unix-like systems a small program \code{xclip} has to be installed, see
References below. Writing is performed via a \code{pipe} with description
\code{xclip -i -selection clipboard}.  By default we are using X selection
\code{clipboard}. Alternatives are \code{XA_PRIMARY} or \code{XA_SECONDARY}.
Consult further documentation of \code{xclip}.

On MacOS writing is performed via \code{pipe} using \code{pbcopy} description
by default.

\code{read_cb} can be called without any arguments. Reading is performed via
a connection description of which can be specified with \code{condesc} or
\code{clipboard.read} option.  Defaults are system specific, see below.

Reading Windows clipboard uses \code{file} connection with description \code{clipboard}.

Reading clipboard on Linux is performed via a \code{pipe} and \code{xclip} program.

Reading clipboard on MacOS uses \code{pipe} with description \code{pbcopy}.
}
\references{
Program \code{xclip} \url{http://sourceforge.net/projects/xclip}
}
\seealso{
Section "Clipboard" of \code{help(connection)}.
}

