\name{write_cb}
\alias{write_cb}
\title{Writing to system clipboard}
\usage{
  write_cb(x, ...)
}
\arguments{
  \item{x}{character, lines of text to be written}

  \item{...}{other arguments passed to/from other methods}
}
\value{
  nothing
}
\description{
  System-independent writing to clipboard for pasting
  elsewhere
}

